+++ src/shared/api.c
@@ -17,6 +17,12 @@ api_init(duk_context* ctx)
 		| DUK_DEFPROP_CLEAR_ENUMERABLE
 		| DUK_DEFPROP_SET_WRITABLE
 		| DUK_DEFPROP_SET_CONFIGURABLE);
+	duk_push_string(ctx, "exports");
+	duk_push_global_object(ctx);
+	duk_def_prop(ctx, -3, DUK_DEFPROP_HAVE_VALUE
+		| DUK_DEFPROP_CLEAR_ENUMERABLE
+		| DUK_DEFPROP_SET_WRITABLE
+		| DUK_DEFPROP_SET_CONFIGURABLE);
 	duk_pop(ctx);
 
 	// set up a prototype stash.  this ensures the prototypes for built-in classes